requirements.txt
asgiref==3.3.4
attrs==20.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2
django-filter==2.4.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
drf-spectacular==0.15.1
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
jsonschema==3.2.0
Markdown==3.3.4
MarkupSafe==1.1.1
oauthlib==3.1.0
packaging==20.9
Pillow==8.2.0
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
